+++ Ref/Top/RefPackets.fppi
@@ -1,4 +1,4 @@
-telemetry packets RefPackets {
+telemetry packets ackets {
 
   packet CDH id 1 group 1 {
     CdhCore.cmdDisp.CommandsDispatched
@@ -9,10 +9,10 @@ telemetry packets RefPackets {
     FileHandling.fileDownlink.PacketsSent
     FileHandling.fileManager.CommandsExecuted
 
-    Ref.cmdSeq.CS_LoadCommands
-    Ref.cmdSeq.CS_CancelCommands
-    Ref.cmdSeq.CS_CommandsExecuted
-    Ref.cmdSeq.CS_SequencesCompleted
+    cmdSeq.CS_LoadCommands
+    cmdSeq.CS_CancelCommands
+    cmdSeq.CS_CommandsExecuted
+    cmdSeq.CS_SequencesCompleted
     ComCcsds.comQueue.comQueueDepth
     ComCcsds.comQueue.buffQueueDepth
     ComCcsds.commsBufferManager.TotalBuffs
@@ -21,9 +21,9 @@ telemetry packets RefPackets {
     #CdhCore.tlmSend.GroupConfigs
     #CdhCore.tlmSend.SectionEnabled
 
-    Ref.rateGroup1Comp.RgMaxTime
-    Ref.rateGroup2Comp.RgMaxTime
-    Ref.rateGroup3Comp.RgMaxTime
+    rateGroup1Comp.RgMaxTime
+    rateGroup2Comp.RgMaxTime
+    rateGroup3Comp.RgMaxTime
   }
 
   packet CDHErrors id 2 group 1 {
@@ -34,151 +34,151 @@ telemetry packets RefPackets {
     FileHandling.fileDownlink.Warnings
     FileHandling.fileManager.Errors
 
-    Ref.cmdSeq.CS_Errors
+    cmdSeq.CS_Errors
     ComCcsds.commsBufferManager.NoBuffs
     ComCcsds.commsBufferManager.EmptyBuffs
 
-    Ref.rateGroup1Comp.RgCycleSlips
-    Ref.rateGroup2Comp.RgCycleSlips
-    Ref.rateGroup3Comp.RgCycleSlips
+    rateGroup1Comp.RgCycleSlips
+    rateGroup2Comp.RgCycleSlips
+    rateGroup3Comp.RgCycleSlips
   }
 
   packet DriveTlm id 3 group 1 {
-    Ref.pingRcvr.PR_NumPings
-    Ref.sendBuffComp.PacketsSent
-    Ref.sendBuffComp.NumErrorsInjected
-    Ref.sendBuffComp.Parameter3
-    Ref.sendBuffComp.Parameter4
-    Ref.sendBuffComp.SendState
-    Ref.recvBuffComp.PktState
-    Ref.recvBuffComp.Sensor1
-    Ref.recvBuffComp.Sensor2
-    Ref.recvBuffComp.Parameter1
-    Ref.recvBuffComp.Parameter2
-    Ref.blockDrv.BD_Cycles
+    pingRcvr.PR_NumPings
+    sendBuffComp.PacketsSent
+    sendBuffComp.NumErrorsInjected
+    sendBuffComp.Parameter3
+    sendBuffComp.Parameter4
+    sendBuffComp.SendState
+    recvBuffComp.PktState
+    recvBuffComp.Sensor1
+    recvBuffComp.Sensor2
+    recvBuffComp.Parameter1
+    recvBuffComp.Parameter2
+    blockDrv.BD_Cycles
   }
 
   packet SigGenSum id 4 group 1 {
-    Ref.SG1.Output
-    Ref.SG1.Type
-    Ref.SG2.Output
-    Ref.SG2.Type
-    Ref.SG3.Output
-    Ref.SG3.Type
-    Ref.SG4.Output
-    Ref.SG4.Type
-    Ref.SG5.Output
-    Ref.SG5.Type
+    SG1.Output
+    SG1.Type
+    SG2.Output
+    SG2.Type
+    SG3.Output
+    SG3.Type
+    SG4.Output
+    SG4.Type
+    SG5.Output
+    SG5.Type
   }
 
   packet SystemRes1 id 5 group 2 {
-    Ref.systemResources.MEMORY_TOTAL
-    Ref.systemResources.MEMORY_USED
-    Ref.systemResources.NON_VOLATILE_TOTAL
-    Ref.systemResources.NON_VOLATILE_FREE
+    systemResources.MEMORY_TOTAL
+    systemResources.MEMORY_USED
+    systemResources.NON_VOLATILE_TOTAL
+    systemResources.NON_VOLATILE_FREE
   }
 
   packet SystemRes3 id 6 group 2 {
-    Ref.systemResources.CPU
-    Ref.systemResources.CPU_00
-    Ref.systemResources.CPU_01
-    Ref.systemResources.CPU_02
-    Ref.systemResources.CPU_03
-    Ref.systemResources.CPU_04
-    Ref.systemResources.CPU_05
-    Ref.systemResources.CPU_06
-    Ref.systemResources.CPU_07
-    Ref.systemResources.CPU_08
-    Ref.systemResources.CPU_09
-    Ref.systemResources.CPU_10
-    Ref.systemResources.CPU_11
-    Ref.systemResources.CPU_12
-    Ref.systemResources.CPU_13
-    Ref.systemResources.CPU_14
-    Ref.systemResources.CPU_15
+    systemResources.CPU
+    systemResources.CPU_00
+    systemResources.CPU_01
+    systemResources.CPU_02
+    systemResources.CPU_03
+    systemResources.CPU_04
+    systemResources.CPU_05
+    systemResources.CPU_06
+    systemResources.CPU_07
+    systemResources.CPU_08
+    systemResources.CPU_09
+    systemResources.CPU_10
+    systemResources.CPU_11
+    systemResources.CPU_12
+    systemResources.CPU_13
+    systemResources.CPU_14
+    systemResources.CPU_15
   }
 
   packet SigGen1Info id 10 group 2 {
-    Ref.SG1.Info
+    SG1.Info
   }
 
   packet SigGen2Info id 11 group 2 {
-    Ref.SG2.Info
+    SG2.Info
   }
 
   packet SigGen3Info id 12 group 2 {
-    Ref.SG3.Info
+    SG3.Info
   }
 
   packet SigGen4Info id 13 group 2 {
-    Ref.SG4.Info
+    SG4.Info
   }
 
   packet SigGen5Info id 14 group 2 {
-    Ref.SG5.Info
+    SG5.Info
   }
 
   packet SigGen1 id 15 group 3 {
-    Ref.SG1.PairOutput
-    Ref.SG1.History
-    Ref.SG1.PairHistory
-    Ref.SG1.DpBytes
-    Ref.SG1.DpRecords
+    SG1.PairOutput
+    SG1.History
+    SG1.PairHistory
+    SG1.DpBytes
+    SG1.DpRecords
   }
 
   packet SigGen2 id 16 group 3 {
-    Ref.SG2.PairOutput
-    Ref.SG2.History
-    Ref.SG2.PairHistory
-    Ref.SG2.DpBytes
-    Ref.SG2.DpRecords
+    SG2.PairOutput
+    SG2.History
+    SG2.PairHistory
+    SG2.DpBytes
+    SG2.DpRecords
   }
 
   packet SigGen3 id 17 group 3 {
-    Ref.SG3.PairOutput
-    Ref.SG3.History
-    Ref.SG3.PairHistory
-    Ref.SG3.DpBytes
-    Ref.SG3.DpRecords
+    SG3.PairOutput
+    SG3.History
+    SG3.PairHistory
+    SG3.DpBytes
+    SG3.DpRecords
   }
 
   packet SigGen4 id 18 group 3 {
-    Ref.SG4.PairOutput
-    Ref.SG4.History
-    Ref.SG4.PairHistory
-    Ref.SG4.DpBytes
-    Ref.SG4.DpRecords
+    SG4.PairOutput
+    SG4.History
+    SG4.PairHistory
+    SG4.DpBytes
+    SG4.DpRecords
   }
 
   packet SigGen5 id 19 group 3 {
-    Ref.SG5.PairOutput
-    Ref.SG5.History
-    Ref.SG5.PairHistory
-    Ref.SG5.DpBytes
-    Ref.SG5.DpRecords
+    SG5.PairOutput
+    SG5.History
+    SG5.PairHistory
+    SG5.DpBytes
+    SG5.DpRecords
   }
 
   packet TypeDemo id 20 group 3 {
-    Ref.typeDemo.ChoiceCh
-    Ref.typeDemo.ChoicesCh
-    Ref.typeDemo.ExtraChoicesCh
-    Ref.typeDemo.ChoicePairCh
-    Ref.typeDemo.ChoiceSlurryCh
-    Ref.typeDemo.Float1Ch
-    Ref.typeDemo.Float2Ch
-    Ref.typeDemo.Float3Ch
-    Ref.typeDemo.FloatSet
-    Ref.typeDemo.ScalarStructCh
-    Ref.typeDemo.ScalarU8Ch
-    Ref.typeDemo.ScalarU16Ch
-    Ref.typeDemo.ScalarU32Ch
-    Ref.typeDemo.ScalarU64Ch
-    Ref.typeDemo.ScalarI8Ch
-    Ref.typeDemo.ScalarI16Ch
-    Ref.typeDemo.ScalarI32Ch
-    Ref.typeDemo.ScalarI64Ch
-    Ref.typeDemo.ScalarF32Ch
-    Ref.typeDemo.ScalarF64Ch
+    typeDemo.ChoiceCh
+    typeDemo.ChoicesCh
+    typeDemo.ExtraChoicesCh
+    typeDemo.ChoicePairCh
+    typeDemo.ChoiceSlurryCh
+    typeDemo.Float1Ch
+    typeDemo.Float2Ch
+    typeDemo.Float3Ch
+    typeDemo.FloatSet
+    typeDemo.ScalarStructCh
+    typeDemo.ScalarU8Ch
+    typeDemo.ScalarU16Ch
+    typeDemo.ScalarU32Ch
+    typeDemo.ScalarU64Ch
+    typeDemo.ScalarI8Ch
+    typeDemo.ScalarI16Ch
+    typeDemo.ScalarI32Ch
+    typeDemo.ScalarI64Ch
+    typeDemo.ScalarF32Ch
+    typeDemo.ScalarF64Ch
   }
 
   packet DataProducts id 21 group 3 {
