+++ Ref/Top/topology.fpp
@@ -12,7 +12,7 @@ module Ref {
 
 
 
-  topology RefTop {
+  topology Ref {
     # ----------------------------------------------------------------------
     # Subtopology imports
     # ----------------------------------------------------------------------
